# src/services/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "GG TODO Backend API",
    "description": "RESTful API for GG TODO - A personal task management application. Built with Node.js, Express, MongoDB, and Firebase Authentication.",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "contact": {
      "name": "GG TODO Support",
      "email": "support@ggtodo.com"
    }
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Basic authentication required for all API routes"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "JWT Bearer token for authenticated routes"
      }
    },
    "schemas": {
...
      "Todo": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Todo's unique identifier"
          },
          "userId": {
            "type": "string",
            "description": "Owner user's ID"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "description": "Todo title"
          },
          "description": {
            "type": "string",
            "minLength": 5,
            "description": "Todo description"
          },
          "priority": {
            "type": "integer",
            "enum": [1, 2, 3, 4],
            "description": "Priority level: 1=Low, 2=Medium, 3=High, 4=Urgent"
          },
          "status": {
            "type": "integer",
            "enum": [1, 2, 3, 4],
            "description": "Status: 1=Pending, 2=In Progress, 3=Completed, 4=Cancelled"
          },
          "labels": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Array of label strings"
          },
          "dueDate": {
            "type": "string",
            "format": "date-time",
            "description": "Due date for the todo"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "servers": [
    {
      "url": "http://localhost:4200",
      "description": "Local Development"
    }
  ],
...
  "paths": {
...
    "/api/v1/todo": {
      "get": {
        "tags": ["Todo"],
        "summary": "List todos",
        "description": "Get a list of todos for the authenticated user",
        "security": [{ "basicAuth": [] }, { "BearerAuth": [] }],
        "parameters": [
          {
            "name": "accessToken",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of todos",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/Todo" }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["Todo"],
        "summary": "Create a new todo",
        "description": "Create a new todo item for the authenticated user",
        "security": [{ "basicAuth": [] }, { "BearerAuth": [] }],
        "parameters": [
          {
            "name": "accessToken",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "description", "dueDate"],
                "properties": {
                  "title": {
                    "type": "string",
                    "maxLength": 200,
                    "example": "Complete project documentation"
                  },
                  "description": {
                    "type": "string",
                    "minLength": 5,
                    "example": "Write comprehensive documentation for the API endpoints"
                  },
                  "priority": {
                    "type": "integer",
                    "enum": [1, 2, 3, 4],
                    "description": "1=Low, 2=Medium, 3=High, 4=Urgent",
                    "example": 2
                  },
                  "labels": {
                    "type": "array",
                    "items": { "type": "string" },
                    "example": ["work", "documentation"]
                  },
                  "dueDate": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2024-12-31T23:59:59Z"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Todo created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/Todo" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/v1/todo/{todoId}": {
      "put": {
        "tags": ["Todo"],
        "summary": "Update a todo",
        "description": "Update an existing todo by its ID",
        "security": [{ "basicAuth": [] }, { "BearerAuth": [] }],
        "parameters": [
          {
            "name": "accessToken",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "todoId",
            "in": "path",
            "required": true,
            "description": "Todo's MongoDB ObjectId",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "minProperties": 1,
                "properties": {
                  "title": {
                    "type": "string",
                    "maxLength": 200
                  },
                  "description": {
                    "type": "string",
                    "minLength": 5
                  },
                  "priority": {
                    "type": "integer",
                    "enum": [1, 2, 3, 4]
                  },
                  "status": {
                    "type": "integer",
                    "enum": [1, 2, 3, 4],
                    "description": "1=Pending, 2=In Progress, 3=Completed, 4=Cancelled"
                  },
                  "labels": {
                    "type": "array",
                    "items": { "type": "string" }
                  },
                  "dueDate": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Todo updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/Todo" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Todo not found"
          }
        }
      },
      "delete": {
        "tags": ["Todo"],
        "summary": "Delete a todo",
        "description": "Delete a todo by its ID",
        "security": [{ "basicAuth": [] }, { "BearerAuth": [] }],
        "parameters": [
          {
            "name": "accessToken",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "todoId",
            "in": "path",
            "required": true,
            "description": "Todo's MongoDB ObjectId",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Todo deleted successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Todo not found"
          }
        }
      }
    }
  }
}